Move logic to serperate files defined by dodal file structure
